requirements.txt
google_api_python_client==2.50.0
keybert==0.5.1
keyphrase_vectorizers==0.0.7
numpy==1.20.3
pandas==1.3.4
python_dateutil==2.8.2
pytube==12.0.0
sentence_transformers==2.2.0
sentencepiece==0.1.96
streamlit==1.9.0
tensorflow==2.8.0
tensorflow_hub==0.12.0
youtube_transcript_api==0.4.4
